pyproject.toml
[tool.poetry]
name = "api-development-course"
version = "0.1.0"
description = ""
authors = ["Mohsen Pirmoradiyan <mohsen@arboair.com>"]
readme = "README.md"
packages = [{ include = "api_development_course" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.2"
SQLAlchemy = "^2.0.29"
uvicorn = "^0.29.0"
requests = "^2.31.0"
python-multipart = "^0.0.9"
psycopg2-binary = "^2.9.9"
email-validator = "^2.1.1"
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
